pocs/skill_gap_analyzer/requirements.txt
anthropic>=0.40.0
streamlit>=1.30.0
plotly>=5.18.0
python-dotenv>=1.0.0
